findBike por keywords y date

+++ ws-app-model/src/main/java/es/udc/ws/bikes/model/bike/Bike.java
@@ -108,7 +108,7 @@ public class Bike {
 		this.creationDate = creationDate;
 	}
 
-	public double getNumberOfRates() {
+	public int getNumberOfRates() {
 		return numberOfRates;
 	}
